+++ src/sign/index.js
@@ -1,4 +1,4 @@
-import { crypto } from '../crypto/index.js'
+import { crypto_ } from '../crypto/index.js'
 /**
  * @param {import('./index.d.ts').crypto__sign__algorithm_T}algorithm
  * @param {CryptoKey}key
@@ -11,7 +11,8 @@ export async function sign(
 	data
 ) {
 	if (typeof data === 'string') {
-		data = new TextEncoder().encode(data)
+		data = new TextEncoder().encode(/** @type {string} */data)
 	}
+	const crypto = await crypto_()
 	return crypto.subtle.sign(algorithm, key, data)
 }